Removed MQ
