+++ Dockerfile
@@ -18,4 +18,4 @@ COPY . .
 HEALTHCHECK --interval=30s --timeout=10s --start-period=1m CMD http_proxy="" https_proxy="" curl --fail http://${HOST-0.0.0.0}:${PORT:-443}/health || exit 1
 
 # Start container
-CMD [ "node", "receive.js" ]
+CMD [ "node", "index.js" ]
